fix: eslint dialog classes
