+++ src/app/core/components/dialog-info/dialog-info.component.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
-
+import { DialogInfoContentComponent } from './dialog-info-content.component';
 @Component({
   selector: 'app-dialog-info',
   templateUrl: './dialog-info.component.html',
@@ -12,13 +12,4 @@ export class DialogInfoComponent {
   openInfoDialog(): void {
     this.dialog.open(DialogInfoContentComponent);
   }
-}
-
-@Component({
-  selector: 'app-dialog-info-content',
-  templateUrl: 'dialog-info-content.html',
-  styleUrls: ['./dialog-info-content.scss'],
-})
-export class DialogInfoContentComponent {
-  constructor(public dialog: MatDialog) {}
 }
