+++ src/app/core/components/dialog-like/dialog-like.component.ts
@@ -1,5 +1,6 @@
-import { Component, Inject } from '@angular/core';
-import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { Component } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { DialogLikeContentComponent } from './dialog-like-content.component';
 
 @Component({
   selector: 'app-dialog-like',
@@ -11,23 +12,5 @@ export class DialogLikeComponent {
 
   openLikeDialog(): void {
     this.dialog.open(DialogLikeContentComponent);
-  }
-}
-
-@Component({
-  selector: 'app-dialog-like-content',
-  templateUrl: 'dialog-like-content.html',
-})
-export class DialogLikeContentComponent {
-  constructor(public dialogRef: MatDialogRef<DialogLikeComponent>, @Inject(MAT_DIALOG_DATA) public data: any) {}
-
-  onClickLike(): any {
-    const urlRecommend = 'https://docs.google.com/forms/d/e/1FAIpQLScGvcVIuXS0cYSoQBSis1maQ0sjovq4tDRVlSzqammRfHMtAw/viewform';
-    window.open(urlRecommend);
-    this.dialogRef.close();
-  }
-
-  onClickLater(): void {
-    this.dialogRef.close();
   }
 }
